[level 1] Title: 크기가 작은 부분문자열, Time: 1.99 ms, Memory: 61.4 MB -BaekjoonHub

+++ 프로그래머스/1/147355. 크기가 작은 부분문자열/크기가 작은 부분문자열.kt
@@ -9,7 +9,7 @@ class Solution {
 //         return answer
         
         return (0..t.length-p.length).filter {
-            t.substring(it, it+p.length).toLong() <= p.toLong()
+            t.substring(it, it+p.length) <= p
         }.size
     }
 }
